🔨 refactor: reply 페이지 편지 내용 컴포넌트로 분리
